Start rejigging things further by creating a resuming-subscription file/class thingy
